requirements.txt
streamlit==1.38.0
pandas==2.2.2
numpy==1.26.4
altair==5.3.0
python-dateutil==2.9.0.post0
SQLAlchemy==2.0.35
pydantic==2.9.2
pytz==2024.1
regex==2024.9.11
# Optional cloud sync (uncomment if you enable Supabase and add code later)
# supabase==2.6.0
pdfplumber==0.11.4
pytesseract==0.3.13
Pillow==10.4.0
